Fix leaky subscriptions in EditRecordViewModel.
Because the EditRecordViewModel's behavior is conditional based on
whether or not the user is creating a new record or editing an existing
record, I've refactored this so that we conditionally alter a single
stream. Since new and existing records use the same function in
subscriptions, it seemed simpler to to maintain a single stream.