Mbtiles: Set up initial mbtiles db classes
This change introduces the scaffolding for a room database that
implements the mbtiles schema. Eventually, we'll read downloaded mbtiles
files and load them into the application by spinning up a room db based
on their contents.

This change also modifies the gradle build file to add room and
room:rx-java as dependencies.