Bug: Remove markers instead of calling map.clear
The removeMarkers method originally called map.clear, which, in addition
to removing map markers, also clears all polylines, overlays, etc. This
is problematic when it comes to rendering overlays. With this change, we
instead iterate through the list of markers, removing each, so that
rendered overlays will be retained on marker cleanup.