Tile: Add tile room classes
This change adds a tile entity, dao, and entity state to the local room
db. We'll use these entries to keep track of what tiles are already
present on a user's device, as well as determine whether a tile is
pending download, downloading, or has failed to download.
TileEntityState is just a mapping from Tile.State to corresponding db
counterpart states--storing states in the db will allow us to resume
downloads upon resumption of a network connection, should the network
fail mid download.