Offline Imagery: Add download tiles menu option
Introduces a navigation menu item for downloading tiles. When the user
selects this menu option we'll render the JSON layer on the map and
begin the tile download flow. Note that this change only contains
updates to the menu UI, it is not yet hooked up to the corresponding
functionality.