Offline Imagery: Add basemap selector UI
This change adds a fragment, view model, layout, and other components
to support a distinct UI for selecting offline imagery. Keeping this UI
separate ensures we can manage the feature independently, as well as
gives users a better experience (the basemap selection map doesn't
render markers, as these are noisy in the context).

This particular commit doesn't implement any interesting functionality.
It is effectively just a shell for the basemap selection UI. As a
minimal first step, we do render the map in this commit.

This UI is in the nav graph, so users can utilize the android navigation
back arrow to return to the main home screen.