Offline Imagery: Add path method to Tile
This change adds a method for computing a tile's file path to the Tile
class. In particular, the implementation of this method should do
whatever is required to compose a valid file path for the tile given
it's ID.

At present we only need to adjust the positioning of the x y and z
coordinates stored in the tile ID to create a valid path for fetching
tiles from our remote mbtile storage.