Offline Imagery: Make extent states semantic
Rather than maintain a basic and fuzzy sense of the state of an extent
(in terms of selections only), I've added semantic states to extents.
This, in my estimation, improves the management of extent selections.

This change also enables us to handle removals of downloaded extents
using the basemapselector UI.