Remove references to ExtentSelector in basemapselector
Removing existing references to the ExtentSelector interface is prerequisite to
removing the BasemapSelectorMapAdapter. We're removing this adapter since it
used the same map as the normal adapter, leading to bugs, and generally
speaking is superfluous and bloats the overall design.

For now, the basemap selector fragment just uses a plain old mapAdapter
instead.