Add layouts and resources for offline imagery manager
This change adds fragment layouts and res values for the offline imagery area
manager. The area manager is a list of downloaded areas. Users can decide to
remove downloaded areas or download new ones.

This change *also* updated the padding on the two line toolbar layout. This
isn't ideal, as this toolbar is used in other contexts, but at present it's the
only way (that I could uncover) to modify the layout of text on the toolbar
(the padding wasn't quite right).