Update AreaEntity to persist bounds. Add type converters.
Now that the model for areas has changed and we no longer use bounds as
identifiers, we'll need to persist bound information in the AreaEntity. I've
added a new field to do so. I've also added type converters from bounds to
strings so that we may query the DB by bounds without having to perform manual
conversions.