Refactor setActionBar to paramterize title display settings
The previous implementation of setActionBar set the action bar title display to
false. However, we call this function to set the action bar for the offline
areas fragment, which needs to display a title. I've refactored the action bar
settig function to take title display setting as an argument, making the method
flexible enough for use across the current contexts in which we need it.