Add Offline Area Selector UI; Delete old Basemap Selector
This change boots the old basemap selector UI in favor of a new Offline
Area selector. There were a couple of bugs and performance issues with
the old selector (namely stemming from reuing the main map instance +
rendering a massive JSON layer on the map) which the new UI (at least
thus far) is free of.

The initial implementation is pretty basic. We can:

- Navigate to the selector from the Offline Areas list.
- Navigate back to the Offline Areas list from the selector.

The selector renders a map and has a download button which will
eventually fire off area downloads but at the moment does nothing.