Only dismiss the project selector when its visible
This is a small fix related to commit b838347. While that commit fixes
the project selector dismissal, it unconditionally tries to dismiss the
selector when the home fragment stops, however, the selector isn't
always rendered in the home fragement. This leads to scenarios in which
the selector doesn't have an associated fragment manager (since we only
associate one when we show the selector). When we try to dismiss the
selector without a fragment manager, the app crashes.

This change adds a check to ensure the selector is actually visible
before we try to dismiss it.