Refactor JSON geometry methods, return the immutable list of tiles
This change moves two checks for appropriate geojson features into the
GeoJsonTile class, and refactors the tile constructor into several
supporting methods.

Any fields that are requried to specify a tile throw exceptions on their
absence--this way we can debug and diagnose issues in the JSON itself,
as well as potentially inform users of issues in their own JSON later.

We also now actually return the list of constructed tiles form the
parser.