Refactor Tile Download Worker to Poll DB
The Tile Download Worker originally accepted a tile id and the tile
download work manager explicitly queued workers to conduct downloads
of a list of tiles (given their ids). Now the tile download worker
instead checks the local data store for pending tiles and handles
downloading them independently -- the manager no longer passes any input
data to the worker.

I've also made liberal use of completables, removed a method from the
worker, and attempted to simplify the interface.

The findAll method in the TileDao now returns a single, as this behavior
seems more appropriate (typically, we won't need to continue to
continually react to the tile DB contents as a flowable allows, and this
simplifies the use of the stream in the worker, for situations that do
call for a flowable, the single may be converted).