Specify an appropriate file mode based on Tile state
Previously, our download code was erroneous—in cases in which we attempt
to resume a partially downloaded tile, we were using MODE_PRIVATE, which
overwrites, instead of appending, the contents of the tile file,
resulting in partial file contents.

We now set the correct mode (Append or Create) based on the state of the
tile.

I've also fixed the value of the range property, which must specify the
unit used (bytes in our case).

I've also added clarificatory comments on our approach to resuming tile
downloads and fixed some formatting errors.

Also fixed a small PMD error.