Add gnd-geojson; update fileUtil to access raw resources
At the moment, we're using an on-device json file to source mbtiles for
download. We have to bundle the file with the application, so I've added
it as a raw resource.

Accessing the file's contents also requires an additional method in the
FileUtil helper, since android's openRawResource returns an input stream
not a File; it's not the most efficient solution, but in order to have
a small delta and make it easier to change this in the future, I've
implemented a method that copies the inputstream to a file corresponding
to the config name--this allows us to make a drop-in replacement--the
alternative requires rewriting the logic of the json parser, which would
need to change yet again as soon as we no longer use a raw resource, so
I felt this was an appropriate approach for now, even though the stream
to file conversion is redundant.