Improve semantic accuracy of GeoJsonExtents
Previously, we returned vertices for GeoJsonExtents in terms of south/west north/east coordinates--this is a slightly inaccurate picture of the world. In actuality, an extent captures the coordinates of a GeoJson polygon, which is described by a series of linear rings. We only care about the exterior ring, and check for inclusion of tiles in a viewport bounds by determining if any of the points that make up the exterior ring are contained within the viewport bounds.

This commit changes names and slightly alters the implementation to make this clearer. Rather than searching for south/west and north/east boundary points, we simply check all points in the linear ring describing the bounds of a tile.