Display locality/country names in area list
Previously, the area list listed the auto-generated ids for each area on the user's device--which isn't very helpful. Instead, we now use reverse Geocoding to get a human readable descriptor for an area based on the lat/lng of its center.

We use a bottom-up approach, as not all addresses have precise locale detail. If the area centers on a known Feature, we render the feature name, otherwise we fallback to the locality, otherwise, the subadministrative area, otherwise the administrative area, and finally, if all the other granularities are unknown, the country name. If we can't find any locality/address for the area, we list it as "unknown area". This hierarchy follows that of Extensible Address Language spec, in reverse order (a Country subsumes and admin area, an admin area a subadmin area, and so forth).