Rename Tile -> TileSource
TileSource better captures what this object represents--some URL and eventually file path that points to a file that contains tileset data that we can render on the map as an overlay. The old name, Tile, suggests that the object represents a single tile, at a single zoom level, while the tileset files we use actually are pyramidal and contain tiles at multiple zoom levels.