Update local data store to manage tile/area cross references
To simplify our handling of offline area deletions, we now keep an association object in the database to track associations between tile sources and offline areas. When deleting an offline area, we can only delete tiles that are unique to the area in question. This cross-table approach simplifies accounting for this, since we can simply fetch tiles along with the areas that depend on them directly from the data store.

This change also includes the area deletion functionality we need in the data store and offline area repository. I've also updated our File utility to handle deleting files relative to the app's file directory.