Use a custom query to update tile area counts
For integrity reasons, we shouldn't use URL as the primary key for tile sources and should use a UID instead. However, this means we can't rely on Room's automatic update/delete methods which perform selections against primary keys. This commit restores the tile source UID as the primary key and adds custom queries for updating area counts and deleting tile sources based on the URL of the tile source.