chore: update db version
When I attempted to run the app, I got an error about the db schema
changing and the version not being updated. I suppose a change got
pushed at some point that didn't have a corresponding db version bump.
This change simply bumps the db version to fix the issue.