fix: show deletion mutations in sync status UI
Before, we cascaded deletions on the foreign keys for
feature/observation mutations, meaning all mutations for a
feature/observation were wiped from the DB, eradicating the user's
edit/sync history.

This change alters that behavior so that we keep mutations around even
after a feature is deleted. Additionally, it removes a redundant call
to deleteFeature when handling mutations (these are handled server side
and should not be finalized until deleted on the server--the local
project receives this update via syncFeatures).

fixes #1059