fix: execute navigations on the main thread
This fixes an issue on startup whereby navigations were executed on the
IO thread (which android doesn't allow). We compose the auth manager
stream with the userRepository. The userRepository returns an observable
that observes on the IO thread; post-composition, the resulting stream
takes the observation thread of the compose argument; so the final
stream was being observed on the IO thread.

The fix is straightforward: set the resulting stream to observe on the
UI thread after the composition is performed.