Model: Refactor Geometry and LOI models
We now decouple the notion of a location of interest's geometry from our
representation of geometrical objects. We have a single LOI type which
contains various geometrical types (point, polygon, etc.).

Note that the new Polygon representation is stricter than our previous
representation. The outershell of a Polygon now *must* form a linear
ring. I've had to update a few tests to reflect this.