model: delete GeoJsonLocationOfInterest
Going forward, we will only model locations of interest with geometries,
and use geojson as a serialization format only. We'll no longer need the
GeoJsonLocationOfInterest class once this work is complete, so I've
removed it now to avoid requiring unnecessary changes to it as a
side-effect of base LOI changes.