Implement our own geometry model
Defines new interfaces and data classes for representing geometrical
forms in ground. These classes should allow us to drop dependencies on
other geometry libraries and implement the minimum functionality we
need.

Note that this commit only adds the new model classes, they are not yet
used in other parts of the codebase. Adoption of these classes in favor
of the external ones will be handled in a separate commit.