fix: set renderer after clustermanager initialization
I cannot claim to fully know *why* but for some reason, while setting
the renderer for our cluster manager works just fine in an init block in
the application, in tests it results calling cluster on the manager
while the map is still null. I'm totally befuddled by this, so we'll
have to resort to the recommended, less constrained and less certain
good old imperative mutation-based approach recommended by the android
docs. Hooray.