refactor: split localdatastore into subcomponents
Breaks down the monolithic datastore into a composition of a few
dedicated data access components for each model object in the
application. This  will enable us to update the implementations
modularly or swap them out for alternatives. It should also help keep
the diffs focused as we incrementally move to coroutines.