refactor: add map feature abstraction
Adds a new class, Feature, that represents rendered model objects on the
map. Features accept a tag type that may hold arbitrary data associated
with a model object, callers should check the type of the tag to
determine how to appropriately interpret the map feature.

This CL also removes uses of MapLocationOfInterest in favor of Feature,
except in the PolygonDrawing components, which will be updated in a
separate PR.