feat: hide Offline BaseMaps menu item when no sources defined
This alters the behavior of the home screen to hide the Basemaps menu item when the active survey has no basemap sources defined (i.e. we now only show the option when it's actually possible to download basemaps). I had to make a few small fixes to accomplish this:

1. Sync locally cached surveys upon activation; before, we were relying only on the Offline data, which lead to discrepancies between the offline and remote.
2. Update local basemap sources on survey sync.
3. Don't overwrite basemap sources on writing surveys to the local db

Fixes #812