Handle unsupported task types.
We now handle unsupported task types when these are received from remote
data sources. When the Job has other, valid tasks, we simply ignore the
invalid task type. For the edge case in which the Job's one and only
task is invalid, we render a special message to the user, urging them to
contact the survey organizer.