Render polygons opaquely when Jobs lack a default style. (#2125)
Previously, if a Job lacked a default style value, we rendered features
with the color value 0. We use ARGB color format, in which this value
corresponds to a completely transparent color. In result, polygons and
points in a survey would not be rendered at all if a survey's job lacked
a proper default style.

We now instead render features with a default, opaque black color when
the job lacks default styling.