SurveySelector - Only render the latest list of surveys
Fixes some UI jitter by restricting our rendering of surveys in the
selector to the most recent list at the time the user opens the
selector. This means the list isn't fully dynamic (if an organizer adds
them to a survey while the selector is open, they'll need to close then
re-open the selector). However, it does fix the rendering and UX issues
that resulted from our previous dynamism, and this seems like an
acceptable tradeoff.

We do re-render the list in the case that the user removes a survey from
offline use.