Handle unsupported task types. (#2124)
* Handle unsupported task types.

We now handle unsupported task types when these are received from remote
data sources. When the Job has other, valid tasks, we simply ignore the
invalid task type. For the edge case in which the Job's one and only
task is invalid, we render a special message to the user, urging them to
contact the survey organizer.

* Make our static analyzers happy

* Remove task not defined hint

We'll hide the data collection button instead, to be tackled in another
PR.

* Small fixes for task handling:

- Ignore unknown task types when setting suggest LOI tasks
- Use a short circuiting return instead of an if

* Handle unknown data collection strategies in job conversion

* format files