SurveySelector - Only render the latest list of surveys (#2127)
* SurveySelector - Only render the latest list of surveys

Fixes some UI jitter by restricting our rendering of surveys in the
selector to the most recent list at the time the user opens the
selector. This means the list isn't fully dynamic (if an organizer adds
them to a survey while the selector is open, they'll need to close then
re-open the selector). However, it does fix the rendering and UX issues
that resulted from our previous dynamism, and this seems like an
acceptable tradeoff.

We do re-render the list in the case that the user removes a survey from
offline use.

* Revert "SurveySelector - Only render the latest list of surveys"

This reverts commit 283e7cdb7b1250715a140fe20b0848bc53cbb003.

* survey selector: prevent jitter on survey activation

We keep a handle on our list updating coroutine Job and simply cancel it
once the survey is activated (we show a spinner, then the user is
immediately taken to the home screen). This requires calling activations
from the fragment, where we can cancel the Job.

* format files

* SurveySelector: Prevent UI jitter

We accomplish this by setting a stable sort order and applying distinctUntilChanged to our flow.

* SurveySelector: update sort.

Also fixes a detekt issue, removes an unused variables, and formats code.

---------

Co-authored-by: Gino Miceli <228050+gino-m@users.noreply.github.com>