UI: Fix date/time task tests
The previous change (setting the task container layout to 0dp) causes espresso actions to fail in our tests on these views. Normally, android computes the height correctly given view constraints, but this doesn't appear to take place in our tests. For now, fix this by forcing the view to have a height of 1, so that the action may be performed.