UI: Constrain task layouts to avoid overlap. (#2582)
* UI: Constrain task layouts to avoid overlap.

Fixes #2516 by constraining our task container to the position of the action button dock. This ensures long multiple choice tasks aren't covered by the action buttons—the scroll region is limited to viewable area above the action button drawer.

* UI: Fix date/time task tests

The previous change (setting the task container layout to 0dp) causes espresso actions to fail in our tests on these views. Normally, android computes the height correctly given view constraints, but this doesn't appear to take place in our tests. For now, fix this by forcing the view to have a height of 1, so that the action may be performed.

* addedd test cases (#2591)

* Create codecov.yml

* Update codecov.yml

* Upgrade versioner plugin (#2595)

* Upgrade versioner plugin

* Fix gitversioner path

* Reenable codecov reports and PR comments (#2593)

* Rename unit test target

* Always generate code coverage report

* Run tests against devDebug

* Update path of report

* Upgrade versioner version

* Upgrade versioner plugin

* Fix gitversioner path

* Add unit tests for AboutFragment (#2587)

---------

Co-authored-by: Akshay Nandwana <akshaynandwana001@gmail.com>
Co-authored-by: Gino Miceli <228050+gino-m@users.noreply.github.com>