SyncStatus: Use new mutations repository upload queue flow; renaming.
This commit uses the newly introduced getUploadQueueFlow() in the MutationRepository to fetch mutations for the sync status UI. In additiion, I have renamed some classes and added some documentation for clarity. In addition, we now handle SubmissionMutations separately from LOI Mutations, allowing us to better inform users. We also no longer check for the authenticated user in the sync view model, since this is done in the mutation repository before the uplaod queue is returned.