Refactor: Move state management out of Composable in HomeScreenFragment
The changes move the `showUserDetailsDialog` and `showSignOutDialog` state variables and related functions outside of the Composable function in `HomeScreenFragment`. This fixes an issue where navigation clicks would stop working after the first time when adding a Compose view dynamically to the fragment. The state is now managed in the fragment class itself. The Composable still renders the dialogs based on these state variables.