PhotoTask: Don't show draft toast on photo capture (#3093)
* PhotoTask: Don't show draft toast on photo capture.

This PR adds a new flag to track whether or not we are restoring the application from an instantaneous photo capture. If so, we do not bother showing the draft resotred toast.

Since the HomeScreenFragment controls the toast display, we unfortunately need to track this state as shared state in the HomeScreenVM, giving access to this state in the PhotoTaskFragment. We might want to consider moving the toast display into the Task fragments themselves to better encapsulate this state in the future.

Fixes #3080

* HomeScreenVM: Rename WAITING_PHOTO_CAPTURE_KEY

Renamed to AWAITING_PHOTO_CAPTURE_KEY for consistency.