Impose a character limit on text data responses. (#3084)
* Impose a character limit on text data responses.

This commit establishes an initial character limit for freeform text responses (including "other" entries for multiple choice questions). We set an initial limit of 255 characters, configurable in Config.kt. This commit also introduces an error message to indicate to users that they have hit the limit.

* TextTask: Add indicator of character limit.

Also updates the error string presented to users on hitting the text entry character limit.

* MultipleChoiceTask: Add character limit label.

Adds a character limit label to freeform text entry "other" options in multiple choice tasks, in accordance with the label we show for TextTasks.

* Text Tasks: Update character limit UI

Accords with Material recommendations on supporting text for text fields.

* Formatting

* TextTask: Add unit test for character limit check.

* MultiChoice: Add test for character limit check.

* Formatting