Miscellaneous UI Tweaks (#3127)
* JobSelectionModal: Update background.

Updates the background of the job selection modal to match current designs.

* Theme: Update secondary button color.

Adds an enum to control coloring of action buttons dependt on a (for now) binary modality of primary or secondary. Updates the color for secondary actions in the job selection modal accordingly.

* Style: Update size for map action buttons.

* Style: Update bottom margin for home action button

* Style: Update ActionButton icon color.

* Style: Add a gradient shim to home fragment.

Adds a subtle linear gradient to the bottom of the home screen to match the latest UX designs.

* UI: Basemap bottom gradient tweaks.

* UI: Tweak appearance of job selector cards.

* Fix CheckCode violations; formatting etc.

* UI: Further adjustment to basemap gradient